feat(zstd): implement FSE-compressed weights for Huffman literals
- Add build_fse_compressed_weights_header() function for Huffman table optimization
- Integrate automatic selection between direct weights and FSE compression
- Support RFC 8878 section 4.2.1 FSE compression for weight arrays
- Optimize header size by choosing most efficient representation
- Add comprehensive tests for FSE weights functionality
- Enhance 4-stream Huffman literals with FSE weight compression

This completes the 'リテラル: 4 ストリーム + Jump_Table（FSE 圧縮 weights 含む）' task
from TASK_LIST.md, enabling more efficient Huffman literals compression.

+++ crates/nxsh_builtins/src/cp.rs
@@ -445,7 +445,7 @@ fn set_file_times(path: &Path, accessed: SystemTime, modified: SystemTime) -> Re
 
         fn to_filetime(t: SystemTime) -> FILETIME {
             let dur = t.duration_since(UNIX_EPOCH).unwrap_or_default();
-            let mut intervals = (dur.as_secs() as u64) * 10_000_000 + (dur.subsec_nanos() as u64) / 100;
+            let mut intervals = dur.as_secs() * 10_000_000 + (dur.subsec_nanos() as u64) / 100;
             intervals += 11644473600u64 * 10_000_000; // Unix -> Windows epoch offset
             FILETIME { dwLowDateTime: intervals as u32, dwHighDateTime: (intervals >> 32) as u32 }
         }
